Fix MariaDB initialization and numeric schema ranges (#4)
* fix: widen collectable graphics IDs in source schema

* fix: widen collectable graphics IDs in database dump

* fix: map collectable graphics IDs as short

* ci: validate the complete MariaDB dump

* fix: restore complete database dump

* fix: widen galaxy gate parts in schema

* fix: widen galaxy gate parts in schema

* fix: map galaxy gate parts as short

* fix: map galaxy gate parts as short

* fix: widen galaxy gate parts in database dump

* fix: widen graphics IDs in source schema

* fix: widen graphics IDs in source schema

* fix: widen graphics IDs in source schema

* fix: widen graphics IDs in source schema

* fix: map graphics IDs as short

* fix: map graphics IDs as short

* fix: map graphics IDs as short

* fix: map graphics IDs as short

* fix: widen all graphics IDs in database dump

+++ Persistence/database/collectables/structure.sql
@@ -6,7 +6,7 @@ CREATE TABLE `collectables`
 (
     `id`   tinyint      NOT NULL AUTO_INCREMENT
         COMMENT 'Primary Key.',
-    `gfx`  tinyint      NOT NULL DEFAULT 0,
+    `gfx`  smallint     NOT NULL DEFAULT 0,
     `type` tinyint      NOT NULL DEFAULT 0
         COMMENT '0 = box, 1 = ore, 2 = beacon, 3 = firework',
     `name` varchar(255) NOT NULL DEFAULT '',
