+++ Persistence/src/main/java/com/kalaazu/persistence/entity/AccountsShipsEntity.java
@@ -42,7 +42,7 @@ public class AccountsShipsEntity {
 
     @Basic
     @Column(name = "gfx", nullable = false)
-    private byte gfx = 0;
+    private short gfx = 0;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "accounts_id", referencedColumnName = "id", nullable = false)
